requirements.txt
Django==1.5
South==0.7.6
distribute==0.6.28
django-crispy-forms==1.2.3
# django-registration==0.8
-e hg+https://bitbucket.org/ubernostrum/django-registration#egg=Package
MySQL-python==1.2.4
pytz==2013b
wsgiref==0.1.2
gunicorn==0.17.2
django-passwords==0.2.0